feat(delete item): check if own item + has permission before delete

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -51,14 +51,22 @@ const Mutations = {
     );
   },
 
-  async deleteItem(parent, args, ctx, info) {
-    const where = { id:args.id };
-    // find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
-    // check if own item (premission)
-    // TO DO
-    // delete
-    return ctx.db.mutation.deleteItem( { where }, info );
+async deleteItem(parent, args, ctx, info) {
+    const where = { id: args.id };
+    // 1. find the item
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
+    // 3. Delete it!
+    return ctx.db.mutation.deleteItem({ where }, info);
   },
 
   async signup(parent, args, ctx, info) {
